# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Biblioteca",
        "description": "Sistema de Gerenciamento de Biblioteca",
        "contact": {
            "name": "Equipe de Suporte",
            "email": "suporte@biblioteca.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor de API da Biblioteca"
        }
    ],
    "paths": {
        "/api/v1/books": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Lista todos os livros",
                "operationId": "8e67fc126a51598255c56d362d50466e",
                "responses": {
                    "200": {
                        "description": "Lista de livros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Book"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Livros"
                ],
                "summary": "Criar novo livro",
                "operationId": "6c5e00fe01c5e356fa95e728c3cfbc98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livro criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/v1/books/{id}": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Obter detalhes de um livro específico",
                "operationId": "22b313962ebfb68fe2f323a96dc9e91e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do livro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Livros"
                ],
                "summary": "Atualizar livro existente",
                "operationId": "471e8ddd754c8f9e34e69f64d1b53eaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livro atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Livros"
                ],
                "summary": "Remover livro",
                "operationId": "e3e33d11e322f13055b10ead759d294c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "title": "Livro",
                "description": "Modelo de livro",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Dom Quixote"
                    },
                    "author": {
                        "type": "string",
                        "example": "Miguel de Cervantes"
                    },
                    "registration_number": {
                        "type": "string",
                        "example": "LIVRO001"
                    },
                    "genre": {
                        "type": "string",
                        "example": "Ficção"
                    },
                    "status": {
                        "type": "string",
                        "example": "disponível"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "BookCreateRequest": {
                "title": "Requisição de Criação de Livro",
                "required": [
                    "title",
                    "author",
                    "registration_number",
                    "genre"
                ],
                "properties": {
                    "title": {
                        "description": "Título do livro",
                        "type": "string",
                        "example": "Dom Quixote"
                    },
                    "author": {
                        "description": "Autor do livro",
                        "type": "string",
                        "example": "Miguel de Cervantes"
                    },
                    "registration_number": {
                        "description": "Número de registro único do livro",
                        "type": "string",
                        "example": "LIVRO001"
                    },
                    "genre": {
                        "description": "Gênero do livro",
                        "type": "string",
                        "example": "Ficção"
                    }
                },
                "type": "object"
            },
            "BookUpdateRequest": {
                "title": "Requisição de Atualização de Livro",
                "properties": {
                    "title": {
                        "description": "Título do livro",
                        "type": "string",
                        "example": "Dom Quixote"
                    },
                    "author": {
                        "description": "Autor do livro",
                        "type": "string",
                        "example": "Miguel de Cervantes"
                    },
                    "registration_number": {
                        "description": "Número de registro único do livro",
                        "type": "string",
                        "example": "LIVRO001"
                    },
                    "genre": {
                        "description": "Gênero do livro",
                        "type": "string",
                        "example": "Ficção"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}